pyproject.toml
[tool.poetry]
name = "lcs"
version = "0.1.0"
description = "Deterministic library-only LLM consensus (PyPI: lcs, import path src.*); no bundled API/CLI, observability helpers optional."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.4.0"
httpx = "^0.27.2"
prometheus-client = "^0.20.0"
structlog = "^24.2.0"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp = "^1.26.0"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-instrumentation-httpx = "^0.47b0"
radon = "^6.0.1"
pycodestyle = "^2.12.1"
pydocstyle = "^6.3.0"
vulture = "^2.11"
bandit = "^1.7.9"
PyYAML = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
respx = "^0.21.1"
ruff = "^0.6.9"
black = "^24.8.0"
mutmut = "^3.0.0"

[tool.mutmut]
paths_to_mutate = ["src"]
tests_dir = ["tests"]
runner = "python -m pytest -q"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
extend-exclude = ["docker"]
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.poetry.group.test.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-asyncio = "^0.24"
pytest-mock = "^3.14"
